RapidRise - 2026: Generative AI & LLM Engineering Assignment/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
numpy>=1.24.0
